requirements (2).txt
streamlit
wordcloud
matplotlib
pillow
numpy
konlpy
